+++ docs/CHANGELOG.md
@@ -7,7 +7,7 @@ Version 3.22.1-eisop1 (June ?, 2022)
 
 **Closed issues:**
 
-typetools#3030.
+typetools#3030, typetools#3236.
 
 
 Version 3.22.1 (June 1, 2022)